move files to features folder
